+++ tests/Jroc.Test262.Tests/language/expressions/assignment/JavaScript/11.13.1-4-6-s.js
@@ -1,6 +1,16 @@
 // Copyright (c) 2012 Ecma International.  All rights reserved.
 // This code is governed by the BSD license found in the LICENSE file.
 
+/*---
+es5id: 11.13.1-4-6-s
+description: >
+    simple assignment throws TypeError if LeftHandSide is a readonly
+    property in strict mode (Function.length)
+flags: [onlyStrict]
+---*/
+
+"use strict";
+
 function Test262Error(message) {
     this.message = message || "";
     this.name = "Test262Error";